# frontend/src/lib/api/openapi.json
{
  "components": {
    "schemas": {
...
      "AtsType": {
        "enum": [
          "GREENHOUSE",
          "LEVER",
          "ASHBY",
          "SMARTRECRUITERS",
          "WORKDAY",
          "LINKEDIN",
          "INDEED",
          "GENERIC",
          "OTHER"
        ],
        "title": "AtsType",
        "type": "string"
      },
...
      "EmailImportRequest": {
        "additionalProperties": false,
        "description": "A job-alert email (forwarded by n8n or pasted). Only job links are extracted.",
        "properties": {
          "received_at": {
            "anyOf": [
              {
                "maxLength": 100,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Received At"
          },
          "sender": {
            "anyOf": [
              {
                "maxLength": 320,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sender"
          },
          "subject": {
            "default": "",
            "maxLength": 1000,
            "title": "Subject",
            "type": "string"
          },
          "text": {
            "description": "HTML or plain-text body",
            "maxLength": 1000000,
            "title": "Text",
            "type": "string"
          }
        },
        "required": [
          "text"
        ],
        "title": "EmailImportRequest",
        "type": "object"
      },
      "EmailImportResult": {
        "properties": {
          "links_found": {
            "title": "Links Found",
            "type": "integer"
          },
          "results": {
            "items": {
              "$ref": "#/components/schemas/JobImportResult"
            },
            "title": "Results",
            "type": "array"
          }
        },
        "required": [
          "links_found",
          "results"
        ],
        "title": "EmailImportResult",
        "type": "object"
      },
...
      "ErrorResponse": {
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorDetail"
          }
        },
        "required": [
          "error"
        ],
        "title": "ErrorResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "JobApplicationRead": {
        "properties": {
          "candidate_id": {
            "format": "uuid",
            "title": "Candidate Id",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/ApplicationStatus"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "id",
          "candidate_id",
          "status",
          "created_at",
          "updated_at"
        ],
        "title": "JobApplicationRead",
        "type": "object"
      },
...
      "JobImportRequest": {
        "additionalProperties": false,
        "description": "A job the user found (e.g. a LinkedIn posting). The URL is stored, never fetched.",
        "properties": {
          "company": {
            "anyOf": [
              {
                "maxLength": 300,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Company"
          },
          "description": {
            "anyOf": [
              {
                "maxLength": 100000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "location": {
            "anyOf": [
              {
                "maxLength": 300,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Location"
          },
          "posted_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Posted At"
          },
          "posted_text": {
            "anyOf": [
              {
                "maxLength": 100,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Relative date as shown, e.g. \"2 hours ago\"",
            "title": "Posted Text"
          },
          "title": {
            "anyOf": [
              {
                "maxLength": 300,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          },
          "url": {
            "maxLength": 2048,
            "minLength": 1,
            "title": "Url",
            "type": "string"
          }
        },
        "required": [
          "url"
        ],
        "title": "JobImportRequest",
        "type": "object"
      },
      "JobImportResult": {
        "properties": {
          "application_created": {
            "title": "Application Created",
            "type": "boolean"
          },
          "created": {
            "title": "Created",
            "type": "boolean"
          },
          "duplicate": {
            "title": "Duplicate",
            "type": "boolean"
          },
          "job": {
            "$ref": "#/components/schemas/JobRead"
          }
        },
        "required": [
          "job",
          "created",
          "duplicate",
          "application_created"
        ],
        "title": "JobImportResult",
        "type": "object"
      },
...
      "JobPage": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/JobRead"
            },
            "title": "Items",
            "type": "array"
          },
          "limit": {
            "title": "Limit",
            "type": "integer"
          },
          "offset": {
            "title": "Offset",
            "type": "integer"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          },
          "window": {
            "$ref": "#/components/schemas/WindowRead"
          }
        },
        "required": [
          "items",
          "total",
          "limit",
          "offset",
          "window"
        ],
        "title": "JobPage",
        "type": "object"
      },
      "JobRead": {
        "properties": {
          "application_status": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ApplicationStatus"
              },
              {
                "type": "null"
              }
            ],
            "description": "Pipeline status for the candidate, when the job is queued"
          },
          "application_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Application Url"
          },
          "ats_type": {
            "$ref": "#/components/schemas/AtsType"
          },
          "canonical_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Canonical Url"
          },
          "company": {
            "title": "Company",
            "type": "string"
          },
          "company_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Company Id"
          },
          "country": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Country"
          },
          "country_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Country Code"
          },
          "discovered_at": {
            "format": "date-time",
            "title": "Discovered At",
            "type": "string"
          },
          "duplicate_count": {
            "default": 0,
            "title": "Duplicate Count",
            "type": "integer"
          },
          "duplicate_of_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Duplicate Of Id"
          },
          "employment_type": {
            "$ref": "#/components/schemas/EmploymentType"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "last_seen_at": {
            "format": "date-time",
            "title": "Last Seen At",
            "type": "string"
          },
          "location": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Location"
          },
          "posted_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Posted At"
          },
          "posting_date_basis": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Posting Date Basis"
          },
          "posting_date_status": {
            "$ref": "#/components/schemas/PostingDateStatus"
          },
          "remote_status": {
            "$ref": "#/components/schemas/RemoteStatus"
          },
          "salary_currency": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Salary Currency"
          },
          "salary_max": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Salary Max"
          },
          "salary_min": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Salary Min"
          },
          "salary_period": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Salary Period"
          },
          "seniority": {
            "$ref": "#/components/schemas/Seniority"
          },
          "source": {
            "title": "Source",
            "type": "string"
          },
          "source_job_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Job Id"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "window_status": {
            "$ref": "#/components/schemas/WindowStatus"
          }
        },
        "required": [
          "id",
          "source",
          "source_job_id",
          "company",
          "company_id",
          "title",
          "location",
          "country",
          "country_code",
          "remote_status",
          "employment_type",
          "seniority",
          "salary_min",
          "salary_max",
          "salary_currency",
          "salary_period",
          "posted_at",
          "posting_date_status",
          "posting_date_basis",
          "window_status",
          "discovered_at",
          "last_seen_at",
          "application_url",
          "canonical_url",
          "ats_type",
          "duplicate_of_id"
        ],
        "title": "JobRead",
        "type": "object"
      },
...
      "JobStats": {
        "properties": {
          "by_source": {
            "additionalProperties": {
              "type": "integer"
            },
            "title": "By Source",
            "type": "object"
          },
          "duplicates": {
            "description": "Duplicate listings linked to a primary record",
            "title": "Duplicates",
            "type": "integer"
          },
          "found_today": {
            "description": "Jobs discovered since midnight (TIMEZONE)",
            "title": "Found Today",
            "type": "integer"
          },
          "in_window": {
            "description": "Jobs posted within JOB_LOOKBACK_HOURS",
            "title": "In Window",
            "type": "integer"
          },
          "last_discovery": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/LastDiscovery"
              },
              {
                "type": "null"
              }
            ]
          },
          "total": {
            "description": "Primary job records (duplicates excluded)",
            "title": "Total",
            "type": "integer"
          },
          "unknown_date": {
            "title": "Unknown Date",
            "type": "integer"
          },
          "window": {
            "$ref": "#/components/schemas/WindowRead"
          }
        },
        "required": [
          "total",
          "found_today",
          "in_window",
          "unknown_date",
          "duplicates",
          "by_source",
          "window",
          "last_discovery"
        ],
        "title": "JobStats",
        "type": "object"
      },
...
      "PostingDateStatus": {
        "enum": [
          "KNOWN",
          "ESTIMATED",
          "UNKNOWN"
        ],
        "title": "PostingDateStatus",
        "type": "string"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "description": "API token (API_AUTH_TOKEN)",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "\nSelf-hosted AI job application agent.\n\n* **Discovery \u2192 analysis \u2192 CV generation \u2192 application preparation** are automated.\n* **Submission** always requires explicit approval by default (`AUTO_SUBMIT=false`).\n* All endpoints except `/health/*` require `Authorization: Bearer <API_AUTH_TOKEN>`.\n",
    "title": "AI Job Application Agent",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/jobs": {
      "get": {
        "operationId": "list_jobs_api_v1_jobs_get",
        "parameters": [
          {
            "in": "query",
            "name": "window",
            "required": false,
            "schema": {
              "default": "in_window",
              "enum": [
                "in_window",
                "out_of_window",
                "all"
              ],
              "title": "Window",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "date_status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/PostingDateStatus"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Date Status"
            }
          },
          {
            "in": "query",
            "name": "source",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 50,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Source"
            }
          },
          {
            "in": "query",
            "name": "country",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "pattern": "^[A-Za-z]{2}$",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Country"
            }
          },
          {
            "in": "query",
            "name": "q",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 200,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Q"
            }
          },
          {
            "in": "query",
            "name": "company_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "uuid",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Company Id"
            }
          },
          {
            "in": "query",
            "name": "include_duplicates",
            "required": false,
            "schema": {
              "default": false,
              "title": "Include Duplicates",
              "type": "boolean"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 200,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobPage"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List jobs (default: posted in the lookback window)",
        "tags": [
          "jobs"
        ]
      }
    },
    "/api/v1/jobs/import": {
      "post": {
        "operationId": "import_job_api_v1_jobs_import_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/JobImportRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobImportResult"
                }
              }
            },
            "description": "The posting was already known"
          },
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobImportResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Invalid or unsafe URL"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Import a job URL (LinkedIn or any public posting); stored, never fetched",
        "tags": [
          "jobs"
        ]
      }
    },
    "/api/v1/jobs/import/email": {
      "post": {
        "operationId": "import_email_api_v1_jobs_import_email_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EmailImportRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EmailImportResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Import the job links of a job-alert email (used by the n8n workflow)",
        "tags": [
          "jobs"
        ]
      }
    },
    "/api/v1/jobs/stats": {
      "get": {
        "operationId": "job_stats_api_v1_jobs_stats_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobStats"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Discovery statistics for the dashboard",
        "tags": [
          "jobs"
        ]
      }
    },
    "/api/v1/jobs/{job_id}": {
      "get": {
        "operationId": "get_job_api_v1_jobs__job_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Job detail with its other listings and pipeline entries",
        "tags": [
          "jobs"
        ]
      }
    },
    "/api/v1/jobs/{job_id}/track": {
      "post": {
        "operationId": "track_job_api_v1_jobs__job_id__track_post",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobApplicationRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Queue a job for the candidate (e.g. a job whose posting date is unknown)",
        "tags": [
          "jobs"
        ]
      }
    },
...
  }
}